requirements.txt
rich>=13.0.0
click>=8.0.0
colorama>=0.4.6
pyyaml>=6.0.0
pymupdf>=1.23.0
python-docx>=1.1.0
ollama>=0.1.0
